Modifs constructeurs et retrait des setters dans le modèle

+++ lille1/src/main/java/usecase/ActionDistributeur.java
@@ -6,7 +6,7 @@ import model.Distributeur;
 public interface ActionDistributeur {
 	Distributeur afficherSolde(Client client);
 
-	Distributeur retrait(Client client, float montant);
+	Distributeur retrait(Client client, double montant);
 
-	Distributeur virement(Client client, float montant, String ibanTo);
+	Distributeur virement(Client client, double montant, String ibanTo);
 }
